refactor: extract findInstanceWithVariable to operateDashboardHelper
Move the grouped-incident instance lookup out of dashboard.spec.ts into a
shared utils helper to keep the spec focused. No behaviour change.
